Rename to .jsx and add vite
