+++ index.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <!-- google search console: -->
+    <meta
+      name="google-site-verification"
+      content="ELp3DMjxwK7A4Wd9jebTMwvGayaGy62drYqlYzSMMZI"
+    />
+    <meta
+      name="google-site-verification"
+      content="ASfhgN33imfcheNyg4MZBn0tZKCnoc8UbEpcR5tQBrQ"
+    />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta name="theme-color" content="#014f98" />
+    <meta
+      name="description"
+      content="Make your own advanced Discord bot with simple, easy-to-use block coding, without any coding knowledge required!"
+    />
+    <meta
+      name="keywords"
+      content="Discord, discord bot creator, block coding, create a discord bot without coding, block coding for discord, make a discord bot without coding, create a discord bot using blocks, blockly for discord, discord bot without coding, visual coding, blockly, Discord, Discord bot, discord bot creator, s4d, scratch for discord, blockly visual coding, bot creator, bot coding, bot maker, bot generator, discord bot maker, discord bot generator, no coding, without coding, create a discord bot without coding, simple, bot, creator, blockly, simplicity, DisFuse github, DisFuse website, discord app generator, app creator, app maker, discord app generator, discord app maker, app, website, javascript, js, less code, no coding skills, no javascript skills, no js coding skills"
+    />
+    <link rel="icon" href="/media/disfuse-clear.png" />
+    <meta name="application-name" content="DisFuse" />
+    <meta
+      property="og:description"
+      content="Create a Discord bot with block coding!"
+    />
+    <meta property="og:image" content="/media/disfuse.png" />
+    <meta property="og:url" content="/" />
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="DisFuse" />
+    <meta http-equiv="Content-Language" content="en" />
+    <link rel="apple-touch-icon" href="/media/disfuse.png" />
+    <meta name="robots" content="index, follow" />
+    <meta name="revisit-after" content="1 day" />
+
+    <title>DisFuse</title>
+  </head>
+
+  <body>
+    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <div id="root"></div>
+    <script type="module" src="/src/index.jsx"></script>
+  </body>
+</html>
