+++ src/Pages/Dashboard/Favorites.jsx
