+++ src/components/WorkshopItem.jsx
