+++ src/components/WorkspaceBar.jsx
